Fix pressure fields for PnPn-2 full restart
Previously:
In PnPn-2 formulation, the pressure field (pr) was **copied** onto the pressure map field (pm1), leaving in each elements zero values that were stored in the restart file and corrupted the pressure field during the next run. (pressure field  was zero on most of the domain)

Now:
In the PnPn formulation - nothing changes
In the PnPn-2 formulation, the pressure (pr) is **interpolated** onto the velocity mesh as for standard output files. It removes the zeros pressure values on the GLL points